requirements.txt
inquirer==3.1.3
pandas==2.0.2
python-dotenv==1.0.0
Requests==2.31.0
Stripe==7.6.0